Convert encoding of OW page to UTF-8
OW site apparently is using non-UTF-8 encoding today. Not sure how it was
before.

So convert it manually to UTF-8. Otherwise, grep will be complaining and
not output proper output, as it will think the file is binary.

Once we port all to Python, we could be smarter, and detect encoding
dynamically based on HTML headers probably.

Closes: https://github.com/openZH/covid_19/issues/157